requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.0
pytz>=2023.3
openai>=1.0.0
